data-reporting-component/requirements.txt
streamlit>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.2
Pillow>=9.5.0